Further runner improvements
